+++ velo-autoconfigure-common/src/main/java/io/github/luminion/velo/core/VeloCoreAutoConfiguration.java
@@ -1,12 +1,16 @@
 package io.github.luminion.velo.core;
 
 import io.github.luminion.velo.VeloProperties;
+import io.github.luminion.velo.idempotent.IdempotentHandler;
+import io.github.luminion.velo.lock.LockHandler;
+import io.github.luminion.velo.ratelimit.RateLimitHandler;
 import io.github.luminion.velo.spi.Fingerprinter;
 import io.github.luminion.velo.spi.JsonProcessorProvider;
 import io.github.luminion.velo.spi.NamingSuffixStrategy;
 import io.github.luminion.velo.spi.fingerprint.SpelFingerprinter;
 import io.github.luminion.velo.spi.provider.DefaultJsonProcessorProvider;
 import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -45,7 +49,10 @@ public class VeloCoreAutoConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    public VeloBannerPrinter veloBannerPrinter(VeloProperties properties) {
-        return new VeloBannerPrinter(properties);
+    public VeloBannerPrinter veloBannerPrinter(VeloProperties properties,
+            ObjectProvider<IdempotentHandler> idempotentHandler,
+            ObjectProvider<RateLimitHandler> rateLimitHandler,
+            ObjectProvider<LockHandler> lockHandler) {
+        return new VeloBannerPrinter(properties, idempotentHandler, rateLimitHandler, lockHandler);
     }
 }