+++ velo-autoconfigure-common/src/main/java/io/github/luminion/velo/ratelimit/VeloRateLimitAutoConfiguration.java
@@ -5,8 +5,6 @@ import io.github.luminion.velo.core.VeloMessageResolver;
 import io.github.luminion.velo.spi.Fingerprinter;
 import io.github.luminion.velo.ratelimit.aspect.RateLimitAspect;
 import org.aspectj.weaver.Advice;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -28,14 +26,11 @@ import org.springframework.context.annotation.Bean;
 @ConditionalOnProperty(prefix = "velo.rate-limit", name = "enabled", havingValue = "true", matchIfMissing = true)
 public class VeloRateLimitAutoConfiguration {
 
-    private static final Logger log = LoggerFactory.getLogger(VeloRateLimitAutoConfiguration.class);
-
     @Bean
     @ConditionalOnMissingBean(RateLimitAspect.class)
     @ConditionalOnBean({Fingerprinter.class, RateLimitHandler.class})
     public RateLimitAspect rateLimitAspect(VeloProperties properties, Fingerprinter fingerprinter,
             RateLimitHandler rateLimitHandler, ObjectProvider<VeloMessageResolver> messageResolver) {
-        log.info("[Velo Starter] RateLimit enabled, backend handler: {}", rateLimitHandler.getClass().getSimpleName());
         RateLimitAspect aspect = new RateLimitAspect(properties.getRateLimit().getPrefix(), fingerprinter,
                 rateLimitHandler, messageResolver.getIfAvailable());
         aspect.setOrder(properties.getAspectOrder().getRateLimit());